+++ core/src/components/item-sliding/item-sliding.tsx
@@ -270,6 +270,13 @@ export class ItemSliding implements ComponentInterface {
   }
 
   private onStart() {
+    /**
+     * We need to query for the ion-item
+     * every time the gesture starts. Developers
+     * may toggle ion-item elements via *ngIf.
+     */
+    this.item = this.el.querySelector('ion-item');
+
     // Prevent scrolling during gesture
     this.disableContentScrollY();
 
@@ -387,16 +394,28 @@ export class ItemSliding implements ComponentInterface {
         ? SlidingState.Start | SlidingState.SwipeStart
         : SlidingState.Start;
     } else {
+      /**
+       * Item sliding cannot be interrupted
+       * while closing the item. If it did,
+       * it would allow the item to get into an
+       * inconsistent state where multiple
+       * items are then open at the same time.
+       */
+      if (this.gesture) {
+        this.gesture.enable(false);
+      }
       this.tmr = setTimeout(() => {
         this.state = SlidingState.Disabled;
         this.tmr = undefined;
+        if (this.gesture) {
+          this.gesture.enable(true);
+        }
       }, 600) as any;
 
       openSlidingItem = undefined;
       style.transform = '';
       return;
     }
-
     style.transform = `translate3d(${-openAmount}px,0,0)`;
     this.ionDrag.emit({
       amount: openAmount,