+++ core/src/components/picker-internal/picker-internal-interfaces.ts
@@ -1,0 +1,9 @@
+export interface PickerInternalChangeEventDetail {
+  inputMode: boolean;
+  inputModeColumn?: HTMLIonPickerColumnInternalElement;
+}
+
+export interface PickerInternalCustomEvent extends CustomEvent {
+  target: HTMLIonPickerInternalElement;
+  detail: PickerInternalChangeEventDetail;
+}