+++ packages/react/src/hooks/useIonModal.ts
@@ -1,4 +1,5 @@
 import { ModalOptions, modalController } from '@ionic/core/components';
+import { useCallback } from 'react';
 
 import { ReactComponentOrElement } from '../models/ReactComponentOrElement';
 
@@ -11,7 +12,10 @@ import { useOverlay } from './useOverlay';
  * @param componentProps The props that will be passed to the component, if required
  * @returns Returns the present and dismiss methods in an array
  */
-export function useIonModal(component: ReactComponentOrElement, componentProps?: any): UseIonModalResult {
+export function useIonModal(
+  component: ReactComponentOrElement,
+  componentProps?: any
+): UseIonModalResult {
   const controller = useOverlay<ModalOptions, HTMLIonModalElement>(
     'IonModal',
     modalController,
@@ -19,14 +23,14 @@ export function useIonModal(component: ReactComponentOrElement, componentProps?:
     componentProps
   );
 
-  function present(options: Omit<ModalOptions, 'component' | 'componentProps'> & HookOverlayOptions = {}) {
-    controller.present(options as any);
-  };
+  const present = useCallback(
+    (options: Omit<ModalOptions, 'component' | 'componentProps'> & HookOverlayOptions = {}) => {
+      controller.present(options as any);
+    },
+    [controller.present]
+  );
 
-  return [
-    present,
-    controller.dismiss
-  ];
+  return [present, controller.dismiss];
 }
 
 export type UseIonModalResult = [